+++ tomni/annotation_manager/main.py
@@ -50,9 +50,7 @@ class AnnotationManager(object):
                     points=[Point(x=p["x"], y=p["y"]) for p in d["points"]],
                 )
             else:
-                raise ValueError(
-                    f"CDF cannot be created. Dict with id {d.get('id', None)} misses type-key with value ellipse or polygon."
-                )
+                raise ValueError(f"CDF cannot be created. Dict with id {d.get('id', None)} misses type-key with value ellipse or polygon.")
             annotations.append(annotation)
 
         return cls(annotations)
@@ -74,15 +72,7 @@ class AnnotationManager(object):
             for i in range(contour.shape[0]):
                 points.append(Point(x=int(contour[i][0]), y=int(contour[i][1])))
 
-            annotations.append(
-                Polygon(
-                    label="",
-                    id=str(uuid.uuid4()),
-                    children=[],
-                    parents=[],
-                    points=points,
-                )
-            )
+            annotations.append(Polygon(label="", id=str(uuid.uuid4()), children=[], parents=[], points=points))
 
         return cls(annotations)
 
@@ -111,15 +101,7 @@ class AnnotationManager(object):
 
         _, labeled_mask = cv2.connectedComponents(mask, connectivity=connectivity)
 
-        padded_mask = cv2.copyMakeBorder(
-            mask,
-            top=1,
-            bottom=1,
-            left=1,
-            right=1,
-            borderType=cv2.BORDER_CONSTANT,
-            value=0,
-        )
+        padded_mask = cv2.copyMakeBorder(mask, top=1, bottom=1, left=1, right=1, borderType=cv2.BORDER_CONSTANT, value=0)
 
         # If bin mask with 0's and 1's is input then canny fails, so multiply by 255 and clip.
         if padded_mask.max() == 1:
@@ -138,7 +120,7 @@ class AnnotationManager(object):
     @classmethod
     def from_labeled_mask(cls, mask: np.ndarray, include_inner_contours: bool = False):
         """Initializes a AnnotationManager object from a labeled mask.
-        A labeled mask contains components indicated by the same pixel values (see example below). 
+        A labeled mask contains components indicated by the same pixel values (see example below).
 
         Example containing two components yielding a AnnotationManager object with two annotations:
         [[0,0,2,1,1],
@@ -153,11 +135,7 @@ class AnnotationManager(object):
             AnnotationManager: A new AnnotationManager object.
         """
         points = labels2listsOfPoints(mask)
-        contours = [
-            positions2contour(point, return_inner_contours=include_inner_contours,)
-            for point in points
-            if len(point) > 0
-        ]
+        contours = [positions2contour(point, return_inner_contours=include_inner_contours) for point in points if len(point) > 0]
         return AnnotationManager.from_contours(contours)
 
     @classmethod
@@ -171,8 +149,7 @@ class AnnotationManager(object):
 
     @annotations.setter
     def annotations(self, other_annotations: List[Annotation]):
-        """I doubt this setter should be allowed to exist.
-        """
+        """I doubt this setter should be allowed to exist."""
 
     def __len__(self) -> int:
         return len(self._annotations)
@@ -187,8 +164,7 @@ class AnnotationManager(object):
         pass
 
     def __contains__(self, other: Annotation):
-        """to check if self.annotations contains other.
-        """
+        """to check if self.annotations contains other."""
         pass
 
     def __iter__(self):
@@ -203,9 +179,7 @@ class AnnotationManager(object):
         else:
             raise StopIteration
 
-    def to_dict(
-        self, decimals: int = 2, mask: np.ndarray = None, min_overlap: float = 0.9
-    ) -> List[Dict]:
+    def to_dict(self, decimals: int = 2, mask: np.ndarray = None, min_overlap: float = 0.9, **kwargs) -> List[Dict]:
         """Transform AM object to a collection of our format.
 
         Args:
@@ -216,20 +190,11 @@ class AnnotationManager(object):
         """
         if mask is not None:
             filtered_annotations = self._annotations.copy()
-            filtered_annotations = [
-                annotation
-                for annotation in filtered_annotations
-                if annotation.is_in_mask(mask, min_overlap)
-            ]
+            filtered_annotations = [annotation for annotation in filtered_annotations if annotation.is_in_mask(mask, min_overlap)]
 
-            return [
-                annotation.to_dict(decimals=decimals)
-                for annotation in filtered_annotations
-            ]
+            return [annotation.to_dict(decimals=decimals) for annotation in filtered_annotations]
 
-        return [
-            annotation.to_dict(decimals=decimals) for annotation in self._annotations
-        ]
+        return [annotation.to_dict(decimals=decimals) for annotation in self._annotations]
 
     def to_contours(self) -> List[np.ndarray]:
         """Transform AM object to a collection of cv2 contours.
@@ -240,20 +205,15 @@ class AnnotationManager(object):
         Returns:
             List[np.ndarray]: Collection of contours as [[[x_0, y_0],..., [x_n, y_n]], ... ,[[x_0, y_0],..., [x_m, y_m]]]
         """
-        if not all(
-            [isinstance(annotation, Polygon) for annotation in self._annotations]
-        ):
+        if not all([isinstance(annotation, Polygon) for annotation in self._annotations]):
             raise ValueError("`to_contours is only supported on polygon-annotations.`")
 
-        contours = [
-            parse_points_to_contour(annotation.points)
-            for annotation in self._annotations
-        ]
+        contours = [parse_points_to_contour(annotation.points) for annotation in self._annotations]
 
         return contours
 
     def to_binary_mask(self, shape: Tuple[int, int]) -> np.ndarray:
-        """Transform an AM object to a binary mask. 
+        """Transform an AM object to a binary mask.
         Annotations can only be polygon or ellipse.
 
         Args:
@@ -269,7 +229,7 @@ class AnnotationManager(object):
         return mask
 
     def to_labeled_mask(self, shape: Tuple[int, int]) -> np.ndarray:
-        """Transform an AM object to a labeled mask. 
+        """Transform an AM object to a labeled mask.
         Annotations can only be polygon or ellipse.
 
         shape (Tuple[int, int]): Shape of the new labeled mask.
@@ -282,9 +242,7 @@ class AnnotationManager(object):
 
         for annotation in self._annotations:
             if isinstance(annotation, Polygon):
-                points = np.array(
-                    [[point.x, point.y] for point in annotation.points], dtype=np.int32
-                )
+                points = np.array([[point.x, point.y] for point in annotation.points], dtype=np.int32)
                 cv2.fillPoly(mask, [points], color=label_color)
             elif isinstance(annotation, Ellipse):
                 cv2.ellipse(
@@ -298,9 +256,7 @@ class AnnotationManager(object):
                     thickness=-1,
                 )
             else:
-                raise TypeError(
-                    "Innapropiate annotation type for `to_labeled_mask`. Supported annotations are ellipse and polygon."
-                )
+                raise TypeError("Innapropiate annotation type for `to_labeled_mask`. Supported annotations are ellipse and polygon.")
 
             # increase color for every annotation.
             label_color += 1
@@ -338,9 +294,7 @@ class AnnotationManager(object):
         """
         pass
 
-    def filter(
-        self, feature: str, min_val: float, max_val: float, inplace: bool = False,
-    ):
+    def filter(self, feature: str, min_val: float, max_val: float, inplace: bool = False):
         """Filter annotations by feature.
 
         Args:
@@ -367,8 +321,7 @@ class AnnotationManager(object):
 
     @classmethod
     def get_circularity_summary(self):
-        """loops the cdf items to get avg, std, min, max.
-        """
+        """loops the cdf items to get avg, std, min, max."""
 
     def get_feature_summaries(self, features: List[str]) -> Dict:
         """Pass a list of features to calculated.
@@ -379,8 +332,4 @@ class AnnotationManager(object):
         for cdf_item in self._cdf_data:
             circularity.append(cdf_item.circularity)
 
-        return {
-            "circularity": {"avg": 1, "std": 1, "min": 1, "max": 1},
-            "...": "...",
-            "feature_n": {"avg_n": 1, "std_n": 1, "min_n": 1, "max_n": 1},
-        }
+        return {"circularity": {"avg": 1, "std": 1, "min": 1, "max": 1}, "...": "...", "feature_n": {"avg_n": 1, "std_n": 1, "min_n": 1, "max_n": 1}}